Move branch related types into $lib/branch/branch.ts
