+++ apps/desktop/src/components/BranchLane.svelte
@@ -2,12 +2,13 @@
 	import FileCard from '$components/FileCard.svelte';
 	import Resizer from '$components/Resizer.svelte';
 	import Stack from '$components/Stack.svelte';
+	import { BranchStack } from '$lib/branches/branch';
 	import { SelectedOwnership } from '$lib/branches/ownership';
 	import { projectLaneCollapsed } from '$lib/config/config';
 	import { Project } from '$lib/project/project';
 	import { FileIdSelection } from '$lib/selection/fileIdSelection';
 	import { SETTINGS, type Settings } from '$lib/settings/userSettings';
-	import { RemoteFile, BranchStack } from '$lib/vbranches/types';
+	import { RemoteFile } from '$lib/vbranches/types';
 	import {
 		getContext,
 		getContextStoreBySymbol,
