+++ apps/desktop/src/components/CommitMessageInput.svelte
@@ -5,6 +5,7 @@
 	import { PromptService } from '$lib/ai/promptService';
 	import { isFailure } from '$lib/ai/result';
 	import { AIService, type DiffInput } from '$lib/ai/service';
+	import { BranchStack } from '$lib/branches/branch';
 	import { SelectedOwnership } from '$lib/branches/ownership';
 	import {
 		projectAiGenEnabled,
@@ -19,7 +20,7 @@
 	import { splitMessage } from '$lib/utils/commitMessage';
 	import { KeyName } from '$lib/utils/hotkeys';
 	import * as toasts from '$lib/utils/toasts';
-	import { BranchStack, DetailedCommit, Commit } from '$lib/vbranches/types';
+	import { DetailedCommit, Commit } from '$lib/vbranches/types';
 	import { getContext, getContextStore } from '@gitbutler/shared/context';
 	import Checkbox from '@gitbutler/ui/Checkbox.svelte';
 	import Icon from '@gitbutler/ui/Icon.svelte';
