+++ apps/desktop/src/components/FileListItemWrapper.svelte
@@ -1,5 +1,6 @@
 <script lang="ts">
 	import FileContextMenu from './FileContextMenu.svelte';
+	import { BranchStack } from '$lib/branches/branch';
 	import { SelectedOwnership } from '$lib/branches/ownership';
 	import { getLocalCommits, getLocalAndRemoteCommits } from '$lib/commits/contexts';
 	import { getCommitStore } from '$lib/commits/contexts';
@@ -9,7 +10,7 @@
 	import { FileIdSelection } from '$lib/selection/fileIdSelection';
 	import { itemsSatisfy } from '$lib/utils/array';
 	import { computeFileStatus } from '$lib/utils/fileStatus';
-	import { BranchStack, type AnyFile, LocalFile } from '$lib/vbranches/types';
+	import { type AnyFile, LocalFile } from '$lib/vbranches/types';
 	import { getContext, maybeGetContextStore } from '@gitbutler/shared/context';
 	import FileListItem from '@gitbutler/ui/file/FileListItem.svelte';
 	import type { Writable } from 'svelte/store';
