+++ apps/desktop/src/components/HeaderMetaSection.svelte
@@ -2,7 +2,7 @@
 	import SeriesRowLabels from './SeriesLabels.svelte';
 	import BranchLaneContextMenu from '$components/BranchLaneContextMenu.svelte';
 	import ContextMenu from '$components/ContextMenu.svelte';
-	import { PatchSeries } from '$lib/vbranches/types';
+	import { PatchSeries } from '$lib/branches/branch';
 	import Button from '@gitbutler/ui/Button.svelte';
 
 	interface Props {
