+++ apps/desktop/src/components/HunkViewer.svelte
@@ -2,13 +2,14 @@
 	import HunkDiff from './HunkDiff.svelte';
 	import HunkContextMenu from '$components/HunkContextMenu.svelte';
 	import LargeDiffMessage from '$components/LargeDiffMessage.svelte';
+	import { BranchStack } from '$lib/branches/branch';
 	import { SelectedOwnership } from '$lib/branches/ownership';
 	import { draggableElement } from '$lib/dragging/draggable';
 	import { HunkDropData } from '$lib/dragging/draggables';
 	import { Project } from '$lib/project/project';
 	import { SETTINGS, type Settings } from '$lib/settings/userSettings';
 	import { type HunkSection } from '$lib/utils/fileSections';
-	import { BranchStack, type Hunk } from '$lib/vbranches/types';
+	import { type Hunk } from '$lib/vbranches/types';
 	import {
 		getContext,
 		getContextStoreBySymbol,
