+++ apps/desktop/src/components/UncommittedChanges.svelte
@@ -3,8 +3,8 @@
 	import CommitDialog from '$components/CommitDialog.svelte';
 	import Dropzones from '$components/Dropzones.svelte';
 	import InfoMessage from '$components/InfoMessage.svelte';
+	import { BranchStack } from '$lib/branches/branch';
 	import { Project } from '$lib/project/project';
-	import { BranchStack } from '$lib/vbranches/types';
 	import { getContext, getContextStore } from '@gitbutler/shared/context';
 	import type { Writable } from 'svelte/store';
