+++ apps/desktop/src/lib/branches/branchController.ts
@@ -4,8 +4,9 @@ import * as toasts from '$lib/utils/toasts';
 import type { PostHogWrapper } from '$lib/analytics/posthog';
 import type { BaseBranchService } from '$lib/baseBranch/baseBranchService';
 import type { BranchListingService } from '$lib/branches/branchListing';
+import type { StackOrder } from './branch';
 import type { VirtualBranchService } from './virtualBranchService';
-import type { BranchPushResult, Hunk, LocalFile, StackOrder } from '../vbranches/types';
+import type { BranchPushResult, Hunk, LocalFile } from '../vbranches/types';
 
 export type CommitIdOrChangeId = { CommitId: string } | { ChangeId: string };
 export type SeriesIntegrationStrategy = 'merge' | 'rebase' | 'hardreset';
