+++ apps/desktop/src/lib/branches/dragActions.ts
@@ -1,6 +1,7 @@
+import { type BranchStack } from './branch';
 import { filesToOwnership } from '$lib/branches/ownership';
 import { CommitDropData, HunkDropData, FileDropData } from '$lib/dragging/draggables';
-import { LocalFile, type BranchStack } from '$lib/vbranches/types';
+import { LocalFile } from '$lib/vbranches/types';
 import type { BranchController } from '$lib/branches/branchController';
 
 class BranchDragActions {
