+++ apps/desktop/src/lib/branches/virtualBranchService.ts
@@ -1,4 +1,6 @@
-import { DependencyError, PatchSeries, BranchStack, VirtualBranches } from '../vbranches/types';
+import { BranchStack, VirtualBranches } from './branch';
+import { PatchSeries } from './branch';
+import { DependencyError } from '../vbranches/types';
 import { invoke, listen } from '$lib/backend/ipc';
 import { showError } from '$lib/notifications/toasts';
 import { plainToInstance } from 'class-transformer';
